src/backend/requirements.txt
# Smart Task Manager 后端依赖
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openai>=1.0.0
httpx>=0.27.0
apscheduler>=3.10.0
python-multipart>=0.0.6
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
